+++ src/Akkling/Extensions.fs
@@ -11,4 +11,18 @@ module Akkling.Extensions
 
 open Akka.Actor
 
+let (|LifecycleEvent|_|) (message: obj) : LifecycleEvent option =
+    match message with
+    | :? LifecycleEvent as e -> Some e
+    | _ -> None
+
+[<Struct>]
+type CombinedEffect (x: Effect, y: Effect) =
+    interface Effect with
+        member this.OnApplied(context : ExtActor<'Message>, message : 'Message) = 
+            x.OnApplied(context, message)
+            y.OnApplied(context, message)
+
+let inline (@) (x: Effect) (y: Effect) : Effect = CombinedEffect(x, y) :> Effect
+
 let implicit (/) (x:ActorPath) (y:string) = ActorPath.op_Division(x, y)