+++ tests/Akkling.Tests/Actors.fs
@@ -15,7 +15,7 @@ open System
 open Xunit
 
 [<Fact>]
-let ``Actor defined by recursive function responds on series of primitive messagess`` () = testDefault <| fun tck -> 
+let ``Actor defined by recursive function responds on series of primitive messagess`` () : unit = testDefault <| fun tck -> 
     let echo = spawn tck "actor" (actorOf2 <| fun mailbox msg -> mailbox.Sender() <! msg)
 
     echo <! 1
@@ -27,7 +27,7 @@ let ``Actor defined by recursive function responds on series of primitive messag
     expectMsg tck 3 |> ignore
 
 [<Fact>]
-let ``Actor defined by recursive function stops on return Stop`` () = testDefault <| fun tck -> 
+let ``Actor defined by recursive function stops on return Stop`` () : unit = testDefault <| fun tck -> 
     let aref = 
         spawn tck "actor"
         <| fun mailbox ->
@@ -55,7 +55,7 @@ let ``Actor defined by recursive function stops on return Stop`` () = testDefaul
     expectNoMsg tck 
 
 [<Fact>]
-let ``Actor defined by recursive function unhandles message on return Unhandled`` () = testDefault <| fun tck ->
+let ``Actor defined by recursive function unhandles message on return Unhandled`` () : unit = testDefault <| fun tck ->
     let aref = 
         spawn tck "actor"
         <| fun mailbox ->